Restructuring
- Remove examples.
- Renames
