--- examples/php/.env
@@ -1,13 +1,0 @@
-DEBUG=1
-
-DOMAIN=admin.dev.ridi.com
-SESSION_DOMAIN=admin.dev.ridi.com
-
-CMS_RPC_URL=http://127.0.0.1
-COUCHBASE_HOST=127.0.0.1
-
-MEMCACHE_HOST=
-MISC_URL=
-ACTIVE_URL=
-HTTP_HOST_LINK=
-SSL_HOST_LINK=
