--- examples/php/composer.json
@@ -1,18 +1,0 @@
-{
-  "name": "ridibooks/cms-bootstrap-sample",
-  "description": "RIDI CMS bootstrap sample project",
-  "type": "project",
-  "license": "MIT",
-  "require": {
-    "illuminate/database": "^5.2",
-    "ridibooks/cms-sdk": "^1.0.0",
-    "silex/silex": "^2.0",
-    "twig/twig": "^2.0.0",
-    "vlucas/phpdotenv": "^2.4"
-  },
-  "autoload": {
-    "psr-4": {
-      "Ridibooks\\Cms\\Example\\": "src"
-    }
-  }
-}
