--- examples/php/view/home.twig
@@ -1,14 +1,0 @@
-{% extends 'index_new.twig' %}
-
-{% block title %}
-  테스트
-{% endblock %}
-
-{% block body %}
-  <p>Hello, {{ name }}!</p>
-
-  <ul>
-    <li><a href="/example/resource1">link1</a></li>
-    <li><a href="/example/resource2">link2</a></li>
-  </ul>
-{% endblock %}
