--- examples/php/view/resource1.twig
@@ -1,9 +1,0 @@
-{% extends 'index_new.twig' %}
-
-{% block title %}
-  테스트 - resource1
-{% endblock %}
-
-{% block body %}
-  This is resource1
-{% endblock %}
