--- examples/php/view/resource2.twig
@@ -1,9 +1,0 @@
-{% extends 'index_new.twig' %}
-
-{% block title %}
-  테스트 - resource2
-{% endblock %}
-
-{% block body %}
-  This is resource2
-{% endblock %}
