--- examples/php/web/index.php
@@ -1,7 +1,0 @@
-<?php
-
-$app = require __DIR__ . '/../src/app.php';
-require __DIR__ . '/../src/controllers.php';
-
-$app->run();
-
