+++ lib/js/lib/Session/CmsSession.js
@@ -1,4 +1,4 @@
-import authorize from '../Auth';
+import authorize from '../authorize';
 
 class CmsSession {
   constructor(sessionId, sdk) {
@@ -62,12 +62,8 @@ class CmsSession {
 
   authorizeUrl(method, checkUrl) {
     const auths = this.getUserMenuAuths();
-    return new Promise((resolve, reject) => {
-      authorize(auths, method, checkUrl).then((allowed) => {
-        resolve(allowed);
-      }).catch((err) => {
-        reject(err);
-      });
+    return new Promise((resolve) => {
+      resolve(authorize(auths, null, checkUrl));
     });
   }
 }
