+++ lib/js/lib/authorize.js
@@ -1,4 +1,3 @@
-
 const allowedUrls = [
   '/admin/book/pa',
   '/me', // 본인 정보 수정
@@ -33,7 +32,7 @@ const isAuthCorrect = (hash, auth) => {
 };
 
 // 해당 URL의 Hash 권한이 있는지 검사한다.
-const hasHashAuth = (userMenuAuths, hash, checkUrl) => {
+const authorize = (userMenuAuths, hash, checkUrl) => {
   if (allowedUrls.includes(checkUrl)) {
     return true;
   }
@@ -53,6 +52,4 @@ const hasHashAuth = (userMenuAuths, hash, checkUrl) => {
     });
 };
 
-export default function (userMenuAuths, method, checkUrl) {
-  return hasHashAuth(userMenuAuths, null, checkUrl);
-}
+export default authorize;
