+++ lib/js/lib/index.js
@@ -1,31 +1,4 @@
-import CmsClient from './CmsClient';
+import CmsSdk from './CmsSdk';
 import CmsSession from './Session/CmsSession';
-import CmsSessionStore from './Session/CmsSessionStore';
-
-const LoginEndpoint = '/login';
-const CouchbaseBucketName = 'session';
-
-class CmsSdk {
-  constructor(options) {
-    this.cmsClient = new CmsClient(options.cmsRpcUrl);
-    this.sessionStore = new CmsSessionStore(options.couchbaseUri, CouchbaseBucketName);
-  }
-
-  getLoginPageUrl(returnUrl) {
-    return `${LoginEndpoint}?return_url=${encodeURIComponent(returnUrl)}`;
-  }
-
-  getMenuService() {
-    return this.cmsClient.adminMenu;
-  }
-
-  getTagService() {
-    return this.cmsClient.adminTag;
-  }
-
-  getUserService() {
-    return this.cmsClient.adminUser;
-  }
-}
 
 export { CmsSdk, CmsSession };
