+++ lib/js/test/authorize.spec.js
@@ -1,8 +1,8 @@
 import { assert } from 'chai';
-import MenuAuth from '../../lib/Auth/MenuAuth';
+import authorize from '../lib/authorize';
 
-describe('MenuAuth', () => {
-  it('tests if MenuAuth includes test url', () => {
+describe('authorize', () => {
+  it('includes test url', () => {
     const userAuths = [
       {
         id: 9,
@@ -15,6 +15,6 @@ describe('MenuAuth', () => {
       },
     ];
     const testUrl = '/super/logs';
-    assert(MenuAuth(userAuths, '', testUrl));
+    assert(authorize(userAuths, '', testUrl));
   });
 });
